* 기존 feed 에서 담당하고 있는 부분 가져옴
