+++ src/main/java/com/example/demo/preperredPerson/service/PreferredPersonService.java
@@ -1,9 +1,11 @@
 package com.example.demo.preperredPerson.service;
 
 
+import com.example.demo.feed.domain.Feed;
 import com.example.demo.feed.domain.FeedRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -11,7 +13,11 @@ public class PreferredPersonService {
 
     private final FeedRepository feedRepository;
 
-
+    @Transactional
+    public void likeFeedByFeedID(Long id) {
+        Feed feed = feedRepository.findFeedByIdAndDeleted(id, false);
+        feed.like();
+    }
 
 
 }
